Added helper function for converting time
