+++ pages/pomodoro.tsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react';
 import type { NextPage } from 'next';
 import styles from '../styles/Pomodoro.module.css';
 import Head from 'next/head';
+import { millisToMinutesAndSeconds, minutesAndSecondsToMillis } from '../helper/convertTime';
 
 const pomodoro: NextPage = () => {
   const [pomodoro, setPomodoro] = useState<boolean>(true);
@@ -13,7 +14,16 @@ const pomodoro: NextPage = () => {
   const [longBreakTime, setLongBreakTime] = useState<any>(900000);
 
   const [started, setStarted] = useState<boolean>(false);
-  let timer: NodeJS.Timeout;
+
+  let pomodoroTimer: number;
+  let shortBreakTimer: number;
+  let longBreakTimer: number;
+
+  useEffect(() => {
+    setPomodoroTime(millisToMinutesAndSeconds(pomodoroTime))
+    setShortBreakTime(millisToMinutesAndSeconds(shortBreakTime))
+    setLongBreakTime(millisToMinutesAndSeconds(longBreakTime))
+  }, [])
 
   useEffect(() => {
     setStarted(false);
@@ -29,42 +39,7 @@ const pomodoro: NextPage = () => {
     }
   }, [pomodoro, shortBreak, longBreak])
 
-  const millisToMinutesAndSeconds = (millis: number): string => {
-    let minutes: number = Math.floor(millis / 60000);
-    let seconds: any = ((millis % 60000) / 1000).toFixed(0);
-    let time: string = `${minutes}:${(seconds < 10 ? "0" : "")}${seconds}`;
-    if (time === '0:00') {
-      clearInterval(timer);
-      window.alert('Time is up!');
-    }
-    return time;
-  }
-
-  const minutesAndSecondsToMillis = (time: string): number => {
-    let timeArr = time.split(':');
-    let minutes: number = 0;
-    let seconds: number = 0;
-    let milliseconds: number = 0;
-    if (timeArr[0].length === 1) {
-      minutes = parseInt(time.slice(0, 1)) * 60000;
-      seconds = parseInt(time.slice(2)) * 1000;
-    } else if (timeArr[0].length === 2) {
-      minutes = parseInt(time.slice(0, 2)) * 60000;
-      seconds = parseInt(time.slice(3)) * 1000;
-    }
-
-    milliseconds = minutes + seconds;
-    return milliseconds;
-  }
-
-  useEffect(() => {
-    setPomodoroTime(millisToMinutesAndSeconds(pomodoroTime))
-    setShortBreakTime(millisToMinutesAndSeconds(shortBreakTime))
-    setLongBreakTime(millisToMinutesAndSeconds(longBreakTime))
-  }, [])
-
-  const linkClickHandler = (target: string) => {
-    clearInterval(timer);
+  const linkClickHandler = (target: string): void => {
     if (target === 'Pomodoro') {
       setPomodoro(true);
       setShortBreak(false);
@@ -82,12 +57,28 @@ const pomodoro: NextPage = () => {
 
   const startClickHandler = (): void => {
     setStarted(true);
-    timer = setInterval(startTimer, 1000);
+    if (pomodoro) {
+      pomodoroTimer = window.setInterval(startTimer, 1000);
+    } else if (shortBreak) {
+      shortBreakTimer = window.setInterval(startTimer, 1000);
+    } else if (longBreak) {
+      longBreakTimer = window.setInterval(startTimer, 1000);
+    }
   }
 
   const stopClickHandler = (): void => {
     setStarted(false);
-    clearInterval(timer);
+    if (pomodoro) {
+      clearInterval(pomodoroTimer);
+      pomodoroTimer = 0;
+    } else if (shortBreak) {
+      clearInterval(shortBreakTimer);
+      shortBreakTimer = 0;
+    } else if (longBreak) {
+      clearInterval(longBreakTimer);
+      longBreakTimer = 0;
+    }
+    window.alert('Timer should stop');
   }
 
   const startTimer = (): void => {
